+++ CHANGELOG.md
@@ -6,10 +6,13 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 
 - Feature: Handle multi-level options in the Select.
 - Feature: Select can now be clearable, searchable, and multi.
+- Fix: disable React on Rails console traces in dev style guide.
 
 ## [v.4.4.0] - 2016-12-27
 
 Features:
+- Add new `Stepper` component.
+- Add new `ArrowIcon` and `CheckedIcon` components.
 - Add `commissionRules` prop to `LoanSimulator`.
 
 Fixes: